# Scheduled events fire order - Performance