--- a/Common/QuantConnect.csproj
+++ b/Common/QuantConnect.csproj
@@ -349,6 +349,8 @@
     <Compile Include="Python\PythonWrapper.cs" />
     <Compile Include="RealTimeProvider.cs" />
     <Compile Include="Scheduling\ScheduledEventException.cs" />
+    <Compile Include="Scheduling\TimeConsumer.cs" />
+    <Compile Include="Scheduling\TimeMonitor.cs" />
     <Compile Include="Securities\BuyingPower.cs" />
     <Compile Include="Securities\BuyingPowerModel.cs" />
     <Compile Include="Securities\BuyingPowerModelExtensions.cs" />