<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Scheduled events will fire in order even after market hours
- Improving scheduled events performance
- Adding regression and unit test

```
No thermal throttling - ScheduledEventsBenchmark 600 events, 2011 to 2018
MASTER
70.20 seconds at 10k data points per second. Processing total of 696,408 data points.
70.41 seconds at 10k data points per second. Processing total of 696,408 data points.
70.41 seconds at 10k data points per second. Processing total of 696,408 data points.
PR
19.09 seconds at 36k data points per second. Processing total of 696,408 data points.
19.08 seconds at 37k data points per second. Processing total of 696,408 data points.
19.07 seconds at 37k data points per second. Processing total of 696,408 data points.
No scheduledEvent
14.30 seconds at 49k data points per second. Processing total of 696,408 data points.
13.46 seconds at 52k data points per second. Processing total of 696,408 data points.
14.07 seconds at 50k data points per second. Processing total of 696,408 data points.
```
#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/4131
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
- ScheduledEvents fire in order
- Improve performance
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit, regression and performance tests, 
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->